refactor: core/ruleEngine→model/analyzeRule, add AnalyzeUrl (mirrors legado)
Match legado Android structure:
  core/ruleEngine/ → model/analyzeRule/  (io.legado.app.model.analyzeRule)
  core/network/AnalyzeUrl.ts → model/analyzeRule/AnalyzeUrl.ts

File renames also mirror legado class names:
  AnalyzeByCSS.ts → AnalyzeByJSoup.ts
  AnalyzeByJSONPath.ts → AnalyzeByJSonPath.ts

Pure rename — no logic changes. Updates all imports.
