+++ src/model/analyzeRule/AnalyzeByRegex.ts
