+++ src/model/analyzeRule/JSEngine.ts
