+++ src/model/analyzeRule/RuleAnalyzer.ts
