+++ src/model/analyzeRule/index.ts
@@ -1,8 +1,8 @@
 export { RuleAnalyzer, parseRule, parseSingleRule, splitByAt } from './RuleAnalyzer';
 export type { ParsedRule, RuleSegment } from './RuleAnalyzer';
-export { AnalyzeByCSS } from './AnalyzeByCSS';
+export { AnalyzeByCSS } from './AnalyzeByJSoup';
 export { AnalyzeByXPath } from './AnalyzeByXPath';
-export { AnalyzeByJSONPath } from './AnalyzeByJSONPath';
+export { AnalyzeByJSONPath } from './AnalyzeByJSonPath';
 export { AnalyzeByRegex } from './AnalyzeByRegex';
 export { JSEngine } from './JSEngine';
 export type { JSContext } from './JSEngine';
